requirements.txt
transformers
langchain
langchain-huggingface
torch
python-dotenv
